# lib/rest/swagger-doc.yaml
...
openapi: 3.0.0
servers:
- url: '/'
info:
  version: 1.0.0
  title: Express Gateway Admin API
  description: Express Gateway Admin API
  contact:
    name: Express Gateway
    url: https://express-gateway.io
    email: api@express-gateway.io
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  '/users':
    get:
      tags:
      - Users
      summary: Returns all users from Express Gateway
      description: Users
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/Users'
    post:
      tags:
      - Users
      summary: Creates new User
      responses:
        '201':
          description: Created
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/User'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateUserRequest'
        description: user to add to the system
        required: true
...
  '/apps':
    get:
      tags:
      - Apps
      summary: Returns all Apps from Express Gateway
      description: Users
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/Apps'
    post:
      tags:
      - Apps
      summary: Creates new App
      responses:
        '201':
          description: Created
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/App'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateAppRequest'
        description: user to add to the system
        required: true
...
  '/credentials':
    post:
      tags:
      - Credentials
      summary: Create Credential
      responses:
        '201':
          description: Created
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/Credential'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateCredentialRequest'
...
  '/scopes':
    get:
      tags:
      - Scopes
      summary: Get all scopes in Express Gateway
      responses:
        '200':
          description: Scopes
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/Scopes'
    post:
      tags:
      - Scopes
      summary: Create scopes in Express Gateway
      responses:
        '201':
          description: Scopes
          content:
            'application/json':
              schema:
                '$ref': '#/components/schemas/Scopes'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/Scopes'
        description: status
...
components:
  schemas:
...
    CreateUserRequest:
      required:
      - username
      - firstname
      - lastname
      properties:
        username:
          type: string
          description: Unique identifier of user entity
          example: steve
        email:
          type: string
          description: email of user, configurable, not required by default
          example: steve@example.com
        firstname:
          type: string
          example: Steve
          description: First name of user (custom property configured in User model,
            required by default)
        lastname:
          type: string
          example: Brown
          description: Last name of user (custom property configured in User model,
            required by default
        redirectUri:
          type: string
          description: Oauth2 specific property, not required by default
          example: http://example.com
      type: object
...
    CreateAppRequest:
      required:
      - name
      - userId
      properties:
        userId:
          type: string
          description: username or user Id of owner
          example: steve
        name:
          type: string
          description: title of the app, this is just a text with NO reference to
            applicationId (property id)
          example: my-app
        redirectUri:
          type: string
          description: Oauth2 specific property, not required by default
          example: http://example.com
      type: object
...
    CreateCredentialRequest:
      required:
      - credential
      - consumerId
      - type
      properties:
        credential:
          type: object
          description: custom properties to include in credential, optional
          properties:
            scopes:
              type: array
              items:
                type: string
        consumerId:
          type: string
        type:
          type: string
          enum:
          - basic-auth
          - key-auth
          - oauth2
      type: object
    Scope:
      description: scope
      type: string
      example: admin
    Scopes:
      description: scopes
      type: object
      properties:
        scopes:
          type: array
          items:
            '$ref': '#/components/schemas/Scope'
    User:
      description: Main item in the system. Typically would be used to represent a
        person in Express Gateway. User can have multiple Applications (Apps) and
        Credentials
      properties:
        username:
          type: string
          description: Unique identifier of user entity, 1-to-1 relation with id,
            both can be used to build references to other objects like apps\credentials
            etc.
          example: steve
        id:
          type: string
          description: Autogenerated Unique identifier of user entity, duplicates
            functionality of username
          example: 47bc9fa2-f245-4b47-9cb4-29b8ccb49728
        email:
          type: string
          description: email of user, configurable, not required by default
          example: steve@example.com
        firstname:
          type: string
          example: Steve
          description: First name of user (custom property configured in User model,
            required by default)
        lastname:
          type: string
          example: Brown
          description: Last name of user (custom property configured in User model,
            required by default)
        isActive:
          type: boolean
          description: Flag showing if user will be able to login
          example: true
        redirectUri:
          type: string
          description: Oauth2 specific property, not required by default
          example: http://example.com
        createdAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Sun Jul 16 2017 00:06:17 GMT+0300 (EEST)
        updatedAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Sun Jul 16 2017 00:06:17 GMT+0300 (EEST)
      type: object
    App:
      description: This entity is designed to represent non human consumers of API
        endpoints, such as mobile application
      properties:
        name:
          description: application name
          example: my-app
          type: string
        id:
          type: string
          description: Autogenerated Unique identifier of an App
          example: 0e13a310-0319-4780-bf66-d10788e08d8a
        userId:
          type: string
          description: Username or UserId that owns the app
          example: 47bc9fa2-f245-4b47-9cb4-29b8ccb49728
        isActive:
          type: boolean
          description: Flag showing if user will be able to login
          example: true
        redirectUri:
          type: string
          description: Oauth2 specific property, not required by default
          example: http://example.com
        createdAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Tue Jul 18 2017 17:04:06 GMT+0300 (EEST)
        updatedAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Tue Jul 18 2017 17:04:06 GMT+0300 (EEST)
      type: object
    Credential:
      description: A container for secret authentication\authorization data
      properties:
        keyId:
          type: string
          description: Only for key-auth, first part of API key
          example: 55tEGsilJkhKoWMS3kkipH
        keySecret:
          type: string
          description: Only for key-auth, second part of API key
          example: 5BNegGCfqW4rhqqCz3A3sM
        consumerId:
          type: string
          description: Username or UserId or AppId that owns the credential
          example: 47bc9fa2-f245-4b47-9cb4-29b8ccb49728
        password:
          type: string
          description: Only for basic-auth, and onfy if autogenerated
          example: 47bc9fa2-f245-4b47-9cb4-29b8ccb49728
        secret:
          type: string
          description: Only for oauth2, and onfy if autogenerated
          example: 47bc9fa2-f245-4b47-9cb4-29b8ccb49728
        type:
          type: string
          enum:
          - basic-auth
          - key-auth
          - oauth2
        isActive:
          type: boolean
          description: Flag showing if user will be able to login
          example: true
        createdAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Tue Jul 18 2017 17:04:06 GMT+0300 (EEST)
        updatedAt:
          type: string
          description: Date and time of User creation, RFC-1123 formatted date stamp
          example: Tue Jul 18 2017 17:04:06 GMT+0300 (EEST)
      type: object
    Users:
      properties:
        users:
          description: Array of Users
          type: array
          items:
            '$ref': '#/components/schemas/User'
        nextKey:
          type: integer
          description: Reserved for future needs, Always 0
      type: object
    Apps:
      properties:
        users:
          description: Array of Apps
          type: array
          items:
            '$ref': '#/components/schemas/App'
        nextKey:
          type: integer
          description: Reserved for future needs, Always 0
      type: object
...